made new tabs component
