--- src/components/Tab.tsx
@@ -1,245 +1,0 @@
-export const Tab = () => {
-  return (
-    <section className="section" id="tabs">
-      <h1 className="title">Tabs</h1>
-      <hr />
-      <div className="tabs">
-        <ul>
-          <li className="is-active">
-            <a>Pictures</a>
-          </li>
-          <li>
-            <a>Music</a>
-          </li>
-          <li>
-            <a>Videos</a>
-          </li>
-          <li>
-            <a>Documents</a>
-          </li>
-        </ul>
-      </div>
-      <br />
-      <div className="tabs is-centered">
-        <ul>
-          <li className="is-active">
-            <a>
-              <span>Pictures</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Music</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Videos</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Documents</span>
-            </a>
-          </li>
-        </ul>
-      </div>
-      <br />
-      <div className="tabs is-small">
-        <ul>
-          <li className="is-active">
-            <a>Pictures</a>
-          </li>
-          <li>
-            <a>Music</a>
-          </li>
-          <li>
-            <a>Videos</a>
-          </li>
-          <li>
-            <a>Documents</a>
-          </li>
-        </ul>
-      </div>
-      <br />
-      <div className="tabs is-medium">
-        <ul>
-          <li className="is-active">
-            <a>Pictures</a>
-          </li>
-          <li>
-            <a>Music</a>
-          </li>
-          <li>
-            <a>Videos</a>
-          </li>
-          <li>
-            <a>Documents</a>
-          </li>
-        </ul>
-      </div>
-      <br />
-      <div className="tabs is-large">
-        <ul>
-          <li className="is-active">
-            <a>Pictures</a>
-          </li>
-          <li>
-            <a>Music</a>
-          </li>
-          <li>
-            <a>Videos</a>
-          </li>
-          <li>
-            <a>Documents</a>
-          </li>
-        </ul>
-      </div>
-      <br />
-      <div className="tabs is-boxed">
-        <ul>
-          <li className="is-active">
-            <a>
-              <span>Pictures</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Music</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Videos</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Documents</span>
-            </a>
-          </li>
-        </ul>
-      </div>
-      <br />
-      <div className="tabs is-toggle">
-        <ul>
-          <li className="is-active">
-            <a>
-              <span>Pictures</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Music</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Videos</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Documents</span>
-            </a>
-          </li>
-        </ul>
-      </div>
-      <br />
-      <div className="tabs is-toggle is-toggle-rounded">
-        <ul>
-          <li className="is-active">
-            <a>
-              <span>Pictures</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Music</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Videos</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Documents</span>
-            </a>
-          </li>
-        </ul>
-      </div>
-      <br />
-      <div className="tabs is-fullwidth">
-        <ul>
-          <li>
-            <a>
-              <span>Left</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Up</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Right</span>
-            </a>
-          </li>
-        </ul>
-      </div>
-      <br />
-      <div className="tabs is-centered is-boxed">
-        <ul>
-          <li className="is-active">
-            <a>
-              <span>Pictures</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Music</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Videos</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Documents</span>
-            </a>
-          </li>
-        </ul>
-      </div>
-      <br />
-      <div className="tabs is-toggle is-fullwidth is-large">
-        <ul>
-          <li className="is-active">
-            <a>
-              <span>Pictures</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Music</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Videos</span>
-            </a>
-          </li>
-          <li>
-            <a>
-              <span>Documents</span>
-            </a>
-          </li>
-        </ul>
-      </div>
-    </section>
-  );
-};
