+++ src/index.ts
@@ -27,8 +27,8 @@ export { Progress } from "./components/Progress";
 export { Radio, RadioList } from "./components/Radio";
 export { Select } from "./components/Select";
 export { Skeleton } from "./components/Skeleton";
-export { Tab } from "./components/Tab";
 export { Table } from "./components/Table";
+export { Tabs } from "./components/Tabs";
 export { TableOfContents } from "./components/TableOfContents";
 export { Tag } from "./components/Tag";
 export { Textarea } from "./components/Textarea";
